+++ packages/components/src/ui/tooltip-button/index.js
@@ -1,2 +1,0 @@
-export { default as TooltipButton } from './component';
-export * from './component';